fix: preserve upstream status on ServiceUnavailableError; relocate proxy test (IGDD-3108)
Address review feedback:
- ServiceUnavailableError now carries the upstream HTTP status (status + response.status) so API routes still propagate a 502/503 to the client instead of defaulting to 500; Fetch/Push pass it through
- move the serverside proxy test under tests/ to match the repo's API-route test layout
